refactor: ♻️ extract detect_pull_requests and helpers to pull.rs (#191)
